requirements.txt
asgiref==3.8.1
astroid==3.3.8
autoflake==2.3.1
autopep8==2.3.2
cfgv==3.4.0
dill==0.3.9
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
django-filter==24.2  # Added for DjangoFilterBackend
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
filelock==3.17.0
flake8==7.1.1
gunicorn==23.0.0
identify==2.6.6
inflection==0.5.1
isort==6.0.0
mccabe==0.7.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pre_commit==4.1.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pylint==3.3.4
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tomlkit==0.13.2
uritemplate==4.1.1
virtualenv==20.29.1
whitenoise==6.9.0
